Updated encryption protocol & bug fixes
I improved the encryption protocol so that it encrypts both the type of value and the data that's stored. I will work on a number encryption system later so that we can encrypt monetary values.

I also did some bug fixing with updateUserInfo and some optimization with the encryption method.